requirements.txt
absl-py==1.2.0
accelerate==0.21.0
aiohttp==3.8.1
aiosignal==1.2.0
alabaster==0.7.12
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
autopep8==1.7.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cryptography @ file:///tmp/build/80754af9/cryptography_1652083738073/work
cycler==0.11.0
datasets==2.4.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dgl-cu113==0.9.0
dglgo==0.0.2
dill==0.3.5.1
docker-pycreds==0.4.0
docutils==0.19
emot==3.1
entrypoints==0.4
evaluate==0.4.0
executing==1.0.0
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.37.1
frozenlist==1.3.1
fsspec==2022.8.2
gensim==4.2.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
huggingface-hub==0.9.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imagesize==1.4.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonpickle==3.0.1
jsonschema==4.15.0
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
littleutils==0.2.2
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.2
multiprocess==0.70.13
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.6
nltk==3.7
notebook==6.4.12
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
numpydoc==1.4.0
oauthlib==3.2.0
ogb==1.3.4
outdated==0.2.1
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
prometheus-client==0.14.1
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.1
pyDeprecate==0.3.2
Pygments==2.13.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil==2.8.2
pytorch-lightning==1.7.4
pytz==2022.2.1
pyvis==0.3.2
PyYAML==6.0
pyzmq==23.2.1
rdkit-pypi==2022.3.5
regex==2022.8.17
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
responses==0.18.0
rouge-score==0.1.2
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smart-open==6.2.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.3.2.post1
Sphinx==5.1.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
stack-data==0.5.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
torch-cluster==1.6.0
torch-geometric==2.1.0.post1
torch-scatter==2.0.9
torch-sparse==0.6.15
torch-spline-conv==1.2.1
torchaudio==0.11.0
torchmetrics==0.9.3
torchvision==0.12.0
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.2
typer==0.6.1
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
wandb==0.15.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
word2vec==0.11.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
